+++ arm-sysregs/src/fake.rs
@@ -12,7 +12,7 @@ use std::sync::Mutex;
 /// type `$type`.
 #[macro_export]
 macro_rules! read_sysreg {
-    ($sysreg:ident $(: $asm_sysreg:ident)?, $type:ty, safe, $fake_sysregs:expr) => {
+    ($sysreg:ident $(: $asm_sysreg:tt)?, $type:ty, safe, $fake_sysregs:expr) => {
         $crate::_paste::paste! {
             #[doc = "Returns the value of the `"]
             #[doc = stringify!($sysreg)]
@@ -22,7 +22,7 @@ macro_rules! read_sysreg {
             }
         }
     };
-    ($(#[$attributes:meta])* $sysreg:ident $(: $asm_sysreg:ident)?, $type:ty, $fake_sysregs:expr) => {
+    ($(#[$attributes:meta])* $sysreg:ident $(: $asm_sysreg:tt)?, $type:ty, $fake_sysregs:expr) => {
         $crate::_paste::paste! {
             #[doc = "Returns the value of the `"]
             #[doc = stringify!($sysreg)]
@@ -33,7 +33,7 @@ macro_rules! read_sysreg {
             }
         }
     };
-    ($sysreg:ident $(: $asm_sysreg:ident)?, $type:ty : $bitflags_type:ty, safe, $fake_sysregs:expr) => {
+    ($sysreg:ident $(: $asm_sysreg:tt)?, $type:ty : $bitflags_type:ty, safe, $fake_sysregs:expr) => {
         $crate::_paste::paste! {
             #[doc = "Returns the value of the `"]
             #[doc = stringify!($sysreg)]
@@ -43,7 +43,7 @@ macro_rules! read_sysreg {
             }
         }
     };
-    ($(#[$attributes:meta])* $sysreg:ident $(: $asm_sysreg:ident)?, $type:ty : $bitflags_type:ty, $fake_sysregs:expr) => {
+    ($(#[$attributes:meta])* $sysreg:ident $(: $asm_sysreg:tt)?, $type:ty : $bitflags_type:ty, $fake_sysregs:expr) => {
         $crate::_paste::paste! {
             #[doc = "Returns the value of the `"]
             #[doc = stringify!($sysreg)]
@@ -60,7 +60,7 @@ macro_rules! read_sysreg {
 /// of type `$type`.
 #[macro_export]
 macro_rules! write_sysreg {
-    ($sysreg:ident $(: $asm_sysreg:ident)?, $type:ty, safe, $fake_sysregs:expr) => {
+    ($sysreg:ident $(: $asm_sysreg:tt)?, $type:ty, safe, $fake_sysregs:expr) => {
         $crate::_paste::paste! {
             #[doc = "Writes `value` to the `"]
             #[doc = stringify!($sysreg)]
@@ -72,7 +72,7 @@ macro_rules! write_sysreg {
     };
     (
         $(#[$attributes:meta])*
-        $sysreg:ident $(: $asm_sysreg:ident)?, $type:ty, $fake_sysregs:expr
+        $sysreg:ident $(: $asm_sysreg:tt)?, $type:ty, $fake_sysregs:expr
     ) => {
         $crate::_paste::paste! {
             #[doc = "Writes `value` to the `"]
@@ -84,7 +84,7 @@ macro_rules! write_sysreg {
             }
         }
     };
-    ($sysreg:ident $(: $asm_sysreg:ident)?, $type:ty : $bitflags_type:ty, safe, $fake_sysregs:expr) => {
+    ($sysreg:ident $(: $asm_sysreg:tt)?, $type:ty : $bitflags_type:ty, safe, $fake_sysregs:expr) => {
         $crate::_paste::paste! {
             #[doc = "Writes `value` to the `"]
             #[doc = stringify!($sysreg)]
@@ -96,7 +96,7 @@ macro_rules! write_sysreg {
     };
     (
         $(#[$attributes:meta])*
-        $sysreg:ident $(: $asm_sysreg:ident)?, $type:ty : $bitflags_type:ty, $fake_sysregs:expr
+        $sysreg:ident $(: $asm_sysreg:tt)?, $type:ty : $bitflags_type:ty, $fake_sysregs:expr
     ) => {
         $crate::_paste::paste! {
             #[doc = "Writes `value` to the `"]